# apps/freeform_usda_meal_analysis_api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freeform USDA Meal Analysis API",
    "description": "\n        USDA FNDDS\u30c7\u30fc\u30bf\u30d9\u30fc\u30b9\u3092\u4f7f\u7528\u3057\u305f\u98df\u4e8b\u5206\u6790API\n\n        ## \u6a5f\u80fd\n        - \u753b\u50cf\u304b\u3089\u98df\u4e8b\u3092\u5206\u6790\u3057\u6804\u990a\u4fa1\u3092\u8a08\u7b97\n        - \u97f3\u58f0\u5165\u529b\u306b\u3088\u308b\u98df\u4e8b\u5206\u6790\n        - \u30ab\u30b9\u30bf\u30de\u30a4\u30ba\u53ef\u80fd\u306aVLM\u30e2\u30c7\u30eb\u8a2d\u5b9a\n        - \u67d4\u8edf\u306a\u30d7\u30ed\u30f3\u30d7\u30c8\u9078\u629e\n\n        ## \u30c7\u30d5\u30a9\u30eb\u30c8\u8a2d\u5b9a\n        - \u30e2\u30c7\u30eb: Qwen3-VL-30B-A3B-Thinking (30B\u30d1\u30e9\u30e1\u30fc\u30bf)\n        - \u30d7\u30ed\u30f3\u30d7\u30c8: v7_experimental\n        - \u691c\u7d22: Full\u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306e\u307f\u4f7f\u7528\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/meal-analyses/complete": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Meal From Image",
        "description": "\u753b\u50cf\u304b\u3089\u98df\u4e8b\u3092\u5206\u6790\u3057\u3066\u6804\u990a\u4fa1\u3092\u8a08\u7b97(Full\u30a4\u30f3\u30c7\u30c3\u30af\u30b9\u306e\u307f\u4f7f\u7528)\n\n## \u30ea\u30af\u30a8\u30b9\u30c8\u30d1\u30e9\u30e1\u30fc\u30bf\n- **image**: \u98df\u4e8b\u753b\u50cf\u30d5\u30a1\u30a4\u30eb(\u5fc5\u9808)\n- **user_context**: \u98df\u4e8b\u306e\u8aac\u660e\u3084\u30b3\u30f3\u30c6\u30ad\u30b9\u30c8(\u30aa\u30d7\u30b7\u30e7\u30f3)\n\n## \u30e2\u30c7\u30eb\u8a2d\u5b9a(\u30aa\u30d7\u30b7\u30e7\u30f3\u3001\u6307\u5b9a\u3057\u306a\u3044\u5834\u5408\u306f\u30c7\u30d5\u30a9\u30eb\u30c8\u5024\u3092\u4f7f\u7528)\n- **model_id**: DeepInfra VLM\u30e2\u30c7\u30ebID\n- **prompt_path**: \u30d7\u30ed\u30f3\u30d7\u30c8\u30d5\u30a1\u30a4\u30eb\u540d(\u4f8b: \"freeform_prompt_usda_format_ver_v7_production_20251027.txt\")\n- **thinking_budget**: \u601d\u8003\u30c8\u30fc\u30af\u30f3\u6570(QVQ\u30e2\u30c7\u30eb\u4f7f\u7528\u6642)\n- **temperature**: \u751f\u6210\u6e29\u5ea6(0.0-2.0)\n- **max_tokens**: \u6700\u5927\u30c8\u30fc\u30af\u30f3\u6570\n\n## \u691c\u7d22\u8a2d\u5b9a(\u30aa\u30d7\u30b7\u30e7\u30f3)\n- **stage1_top_k**: Stage1\u3067\u53d6\u5f97\u3059\u308b\u5019\u88dc\u6570(\u30c7\u30d5\u30a9\u30eb\u30c8: 50)\n- **bm25_weight**: BM25\u691c\u7d22\u306e\u91cd\u307f(\u30c7\u30d5\u30a9\u30eb\u30c8: 0.4)\n- **vector_weight**: Vector\u691c\u7d22\u306e\u91cd\u307f(\u30c7\u30d5\u30a9\u30eb\u30c8: 0.6)\n- **rrf_k**: RRF\u306ek\u30d1\u30e9\u30e1\u30fc\u30bf(\u30c7\u30d5\u30a9\u30eb\u30c8: 60)\n- **reranker_model**: Reranker\u30e2\u30c7\u30eb\u540d(\u30c7\u30d5\u30a9\u30eb\u30c8: Qwen/Qwen3-Reranker-8B)\n- **reranker_instruction**: Reranker\u7528instruction\n- **reranker_top_n**: Reranker\u3067\u8fd4\u3059\u7d50\u679c\u6570\n\n## \u30ec\u30b9\u30dd\u30f3\u30b9\n- \u691c\u51fa\u3055\u308c\u305f\u6599\u7406\u3068\u98df\u6750\u306e\u8a73\u7d30\n- \u6804\u990a\u4fa1\u8a08\u7b97\u7d50\u679c\n- \u51e6\u7406\u6642\u9593\u3068\u30e1\u30bf\u30c7\u30fc\u30bf",
        "operationId": "analyze_meal_from_image_api_v1_meal_analyses_complete_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_meal_from_image_api_v1_meal_analyses_complete_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id",
            "description": "\u5206\u6790ID",
            "example": "a1b2c3d4"
          },
          "input_type": {
            "type": "string",
            "title": "Input Type",
            "description": "\u5165\u529b\u30bf\u30a4\u30d7",
            "example": "image"
          },
          "meal_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Title",
            "description": "\u98df\u4e8b\u5168\u4f53\u306e\u30bf\u30a4\u30c8\u30eb\uff08MAX 30\u6587\u5b57\uff09",
            "example": "Grilled Chicken with Salad"
          },
          "total_dishes": {
            "type": "integer",
            "title": "Total Dishes",
            "description": "\u691c\u51fa\u3055\u308c\u305f\u6599\u7406\u6570",
            "example": 3
          },
          "total_ingredients": {
            "type": "integer",
            "title": "Total Ingredients",
            "description": "\u7dcf\u98df\u6750\u6570",
            "example": 9
          },
          "processing_time_seconds": {
            "type": "number",
            "title": "Processing Time Seconds",
            "description": "\u51e6\u7406\u6642\u9593\uff08\u79d2\uff09",
            "example": 15.65
          },
          "dishes": {
            "items": {
              "$ref": "#/components/schemas/DishDetail"
            },
            "type": "array",
            "title": "Dishes",
            "description": "\u691c\u51fa\u3055\u308c\u305f\u6599\u7406\u4e00\u89a7"
          },
          "total_nutrition": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NutritionInfo"
              }
            ],
            "description": "\u7dcf\u6804\u990a\u4fa1"
          },
          "ai_model_used": {
            "type": "string",
            "title": "Ai Model Used",
            "description": "\u4f7f\u7528AI\u30e2\u30c7\u30eb",
            "example": "Qwen/Qwen2-VL-72B-Instruct"
          },
          "prompt_file_used": {
            "type": "string",
            "title": "Prompt File Used",
            "description": "\u4f7f\u7528\u30d7\u30ed\u30f3\u30d7\u30c8\u30d5\u30a1\u30a4\u30eb",
            "example": "freeform_prompt_usda_format_ver_v7_experimental_20251027.txt"
          },
          "match_rate_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Match Rate Percent",
            "description": "\u6804\u990a\u691c\u7d22\u30de\u30c3\u30c1\u7387\uff08%\uff09",
            "example": 100.0
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UsageInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Token\u4f7f\u7528\u91cf\u3068\u30b3\u30b9\u30c8\u60c5\u5831"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript",
            "description": "\u97f3\u58f0\u8a8d\u8b58\u30c6\u30ad\u30b9\u30c8\uff08\u97f3\u58f0\u5165\u529b\u6642\u306e\u307f\uff09"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "\u8b66\u544a\u30e1\u30c3\u30bb\u30fc\u30b8",
            "example": []
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "input_type",
          "total_dishes",
          "total_ingredients",
          "processing_time_seconds",
          "dishes",
          "total_nutrition",
          "ai_model_used",
          "prompt_file_used",
          "match_rate_percent"
        ],
        "title": "AnalysisResponse",
        "description": "\u5206\u6790\u30ec\u30b9\u30dd\u30f3\u30b9",
        "example": {
          "ai_model_used": "Qwen/Qwen2-VL-72B-Instruct",
          "analysis_id": "a1b2c3d4",
          "dishes": [
            {
              "calculation_metadata": {
                "calculation_method": "weight_based_scaling",
                "ingredient_count": 1,
                "total_weight_g": 150.0
              },
              "ingredients": [
                {
                  "calculated_nutrition": {
                    "calories": 247.5,
                    "carbs": 0.0,
                    "fat": 5.4,
                    "fiber": 0.0,
                    "protein": 46.5,
                    "sodium": 111.0,
                    "sugar": 0.0
                  },
                  "calculation_notes": [
                    "Scaled from 100g base data using factor 1.500"
                  ],
                  "debug_info": {
                    "reranker_results": [
                      "167782",
                      "167783"
                    ],
                    "retriever_candidates": [
                      "167782",
                      "167783",
                      "167784"
                    ],
                    "retry_count": 0
                  },
                  "fdc_id": "167782",
                  "ingredient_name": "Chicken breast, grilled",
                  "nutrition_per_100g": {
                    "calories": 165.0,
                    "carbs": 0.0,
                    "fat": 3.6,
                    "fiber": 0.0,
                    "protein": 31.0,
                    "sodium": 74.0,
                    "sugar": 0.0
                  },
                  "source_db": "usda_fndds",
                  "weight_g": 150.0
                }
              ],
              "total_nutrition": {
                "calories": 247.5,
                "carbs": 0.0,
                "fat": 5.4,
                "fiber": 0.0,
                "protein": 46.5,
                "sodium": 111.0,
                "sugar": 0.0
              }
            },
            {
              "calculation_metadata": {
                "calculation_method": "weight_based_scaling",
                "ingredient_count": 4,
                "total_weight_g": 155.0
              },
              "ingredients": [
                {
                  "calculated_nutrition": {
                    "calories": 17.0,
                    "carbs": 3.3,
                    "fat": 0.3,
                    "fiber": 2.1,
                    "protein": 1.2,
                    "sodium": 8.0,
                    "sugar": 1.2
                  },
                  "calculation_notes": [
                    "Scaled from 100g base data using factor 1.000"
                  ],
                  "fdc_id": "169248",
                  "ingredient_name": "Romaine lettuce, raw",
                  "nutrition_per_100g": {
                    "calories": 17.0,
                    "carbs": 3.3,
                    "fat": 0.3,
                    "fiber": 2.1,
                    "protein": 1.2,
                    "sodium": 8.0,
                    "sugar": 1.2
                  },
                  "source_db": "usda_fndds",
                  "weight_g": 100.0
                },
                {
                  "calculated_nutrition": {
                    "calories": 141.0,
                    "carbs": 2.1,
                    "fat": 14.4,
                    "fiber": 0.0,
                    "protein": 0.75,
                    "sodium": 270.0,
                    "sugar": 1.5
                  },
                  "calculation_notes": [
                    "Scaled from 100g base data using factor 0.300"
                  ],
                  "fdc_id": "173306",
                  "ingredient_name": "Caesar dressing",
                  "nutrition_per_100g": {
                    "calories": 470.0,
                    "carbs": 7.0,
                    "fat": 48.0,
                    "fiber": 0.0,
                    "protein": 2.5,
                    "sodium": 900.0,
                    "sugar": 5.0
                  },
                  "source_db": "usda_fndds",
                  "weight_g": 30.0
                },
                {
                  "calculated_nutrition": {
                    "calories": 39.2,
                    "carbs": 0.32,
                    "fat": 2.58,
                    "fiber": 0.0,
                    "protein": 3.58,
                    "sodium": 152.9,
                    "sugar": 0.09
                  },
                  "calculation_notes": [
                    "Scaled from 100g base data using factor 0.100"
                  ],
                  "fdc_id": "171287",
                  "ingredient_name": "Parmesan cheese, grated",
                  "nutrition_per_100g": {
                    "calories": 392.0,
                    "carbs": 3.2,
                    "fat": 25.8,
                    "fiber": 0.0,
                    "protein": 35.8,
                    "sodium": 1529.0,
                    "sugar": 0.9
                  },
                  "source_db": "usda_fndds",
                  "weight_g": 10.0
                },
                {
                  "calculated_nutrition": {
                    "calories": 61.05,
                    "carbs": 11.025,
                    "fat": 0.99,
                    "fiber": 0.75,
                    "protein": 1.785,
                    "sodium": 104.7,
                    "sugar": 0.87
                  },
                  "calculation_notes": [
                    "Scaled from 100g base data using factor 0.150"
                  ],
                  "fdc_id": "168042",
                  "ingredient_name": "Croutons, plain",
                  "nutrition_per_100g": {
                    "calories": 407.0,
                    "carbs": 73.5,
                    "fat": 6.6,
                    "fiber": 5.0,
                    "protein": 11.9,
                    "sodium": 698.0,
                    "sugar": 5.8
                  },
                  "source_db": "usda_fndds",
                  "weight_g": 15.0
                }
              ],
              "total_nutrition": {
                "calories": 258.25,
                "carbs": 16.745,
                "fat": 18.27,
                "fiber": 2.85,
                "protein": 7.315,
                "sodium": 535.6,
                "sugar": 3.66
              }
            }
          ],
          "input_type": "image",
          "match_rate_percent": 100.0,
          "meal_title": "Chicken & Caesar Salad",
          "processing_time_seconds": 12.34,
          "prompt_file_used": "freeform_prompt_usda_format_ver_v7_production_20251027.txt",
          "total_dishes": 2,
          "total_ingredients": 5,
          "total_nutrition": {
            "calories": 505.75,
            "carbs": 16.745,
            "fat": 23.67,
            "fiber": 2.85,
            "protein": 53.815,
            "sodium": 646.6,
            "sugar": 3.66
          },
          "usage": {
            "estimated_cost_usd": 0.0234,
            "input_tokens": 2500,
            "output_tokens": 450,
            "raw_vlm_output": "<think>\nImage shows grilled chicken breast and a Caesar salad...\n</think>\n\n{\n  \"dishes\": [\n    {\"dish_name\": \"Grilled Chicken Breast\", ...}\n  ]\n}",
            "thinking_tokens": 3200,
            "total_tokens": 6150
          },
          "warnings": []
        }
      },
      "Body_analyze_meal_from_image_api_v1_meal_analyses_complete_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "\u98df\u4e8b\u753b\u50cf\u30d5\u30a1\u30a4\u30eb"
          },
          "user_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Context",
            "description": "\u30e6\u30fc\u30b6\u30fc\u30b3\u30f3\u30c6\u30ad\u30b9\u30c8"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id",
            "description": "VLM\u30e2\u30c7\u30ebID"
          },
          "prompt_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Path",
            "description": "\u30d7\u30ed\u30f3\u30d7\u30c8\u30d5\u30a1\u30a4\u30eb\u30d1\u30b9(prompts/\u4ee5\u4e0b)"
          },
          "prompt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Text",
            "description": "\u30d7\u30ed\u30f3\u30d7\u30c8\u30c6\u30ad\u30b9\u30c8\u5168\u6587(prompt_path\u3088\u308a\u512a\u5148)"
          },
          "thinking_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Budget",
            "description": "\u601d\u8003\u30c8\u30fc\u30af\u30f3\u6570(QVQ\u30e2\u30c7\u30eb\u7528)"
          },
          "enable_thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Thinking",
            "description": "Thinking mode\u306eon/off(Alibaba\u30e2\u30c7\u30eb\u7528)"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "\u751f\u6210\u6e29\u5ea6"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "\u6700\u5927\u30c8\u30fc\u30af\u30f3\u6570"
          },
          "stage1_top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage1 Top K",
            "description": "Stage1\u5019\u88dc\u6570"
          },
          "bm25_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bm25 Weight",
            "description": "BM25\u691c\u7d22\u306e\u91cd\u307f(0.0-1.0)"
          },
          "vector_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Weight",
            "description": "Vector\u691c\u7d22\u306e\u91cd\u307f(0.0-1.0)"
          },
          "rrf_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rrf K",
            "description": "RRF\u306ek\u30d1\u30e9\u30e1\u30fc\u30bf"
          },
          "reranker_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Model",
            "description": "Reranker\u30e2\u30c7\u30eb\u540d"
          },
          "reranker_instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Instruction",
            "description": "Reranker instruction"
          },
          "reranker_top_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Top N",
            "description": "Reranker\u8fd4\u5374\u6570"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_analyze_meal_from_image_api_v1_meal_analyses_complete_post"
      },
      "DishDetail": {
        "properties": {
          "dish_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dish Name",
            "description": "\u6599\u7406\u540d",
            "example": "Grilled Chicken Salad"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientDetail"
            },
            "type": "array",
            "title": "Ingredients",
            "description": "\u98df\u6750\u8a73\u7d30\u30ea\u30b9\u30c8"
          },
          "total_nutrition": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NutritionInfo"
              }
            ],
            "description": "\u6599\u7406\u306e\u7dcf\u6804\u990a\u4fa1"
          },
          "calculation_metadata": {
            "type": "object",
            "title": "Calculation Metadata",
            "description": "\u8a08\u7b97\u30e1\u30bf\u30c7\u30fc\u30bf",
            "example": {
              "calculation_method": "weight_based_scaling",
              "ingredient_count": 2,
              "total_weight_g": 200.0
            }
          }
        },
        "type": "object",
        "required": [
          "ingredients",
          "total_nutrition",
          "calculation_metadata"
        ],
        "title": "DishDetail",
        "description": "\u6599\u7406\u8a73\u7d30\u60c5\u5831"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NutritionInfo": {
        "properties": {
          "calories": {
            "type": "number",
            "title": "Calories",
            "description": "\u30ab\u30ed\u30ea\u30fc\uff08kcal\uff09",
            "example": 156.0
          },
          "protein": {
            "type": "number",
            "title": "Protein",
            "description": "\u30bf\u30f3\u30d1\u30af\u8cea\uff08g\uff09",
            "example": 12.0
          },
          "fat": {
            "type": "number",
            "title": "Fat",
            "description": "\u8102\u8cea\uff08g\uff09",
            "example": 10.5
          },
          "carbs": {
            "type": "number",
            "title": "Carbs",
            "description": "\u70ad\u6c34\u5316\u7269\uff08g\uff09",
            "example": 1.1
          },
          "fiber": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiber",
            "description": "\u98df\u7269\u7e4a\u7dad\uff08g\uff09",
            "example": 0.0
          },
          "sugar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sugar",
            "description": "\u7cd6\u8cea\uff08g\uff09",
            "example": 0.7
          },
          "sodium": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sodium",
            "description": "\u30ca\u30c8\u30ea\u30a6\u30e0\uff08mg\uff09",
            "example": 142.0
          }
        },
        "type": "object",
        "required": [
          "calories",
          "protein",
          "fat",
          "carbs"
        ],
        "title": "NutritionInfo",
        "description": "\u6804\u990a\u4fa1\u60c5\u5831"
      },
...
      "UsageInfo": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "description": "\u5165\u529b\u30c8\u30fc\u30af\u30f3\u6570",
            "example": 1250
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "description": "\u51fa\u529b\u30c8\u30fc\u30af\u30f3\u6570",
            "example": 450
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "\u5408\u8a08\u30c8\u30fc\u30af\u30f3\u6570",
            "example": 1700
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd",
            "description": "\u63a8\u5b9a\u30b3\u30b9\u30c8\uff08USD\uff09(pricing\u60c5\u5831\u304c\u306a\u3044\u5834\u5408\u306fNone)",
            "example": 0.00125
          },
          "model_pricing": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Pricing",
            "description": "\u4f7f\u7528\u3057\u305f\u30e2\u30c7\u30eb\u306e\u4fa1\u683c\u60c5\u5831 (pricing\u60c5\u5831\u304c\u306a\u3044\u5834\u5408\u306fNone)",
            "example": {
              "input_price_per_million": 0.29,
              "output_price_per_million": 0.99
            }
          },
          "raw_vlm_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Vlm Output",
            "description": "VLM\u306e\u751f\u51fa\u529b\uff08\u30c7\u30d0\u30c3\u30b0\u7528\uff09",
            "example": "<think>...</think>\n{\"dishes\": ...}"
          },
          "prompt_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Content",
            "description": "VLM\u306b\u9001\u4fe1\u3057\u305f\u30d7\u30ed\u30f3\u30d7\u30c8\u306e\u5185\u5bb9\uff08\u30c7\u30d0\u30c3\u30b0\u7528\uff09",
            "example": "You are an expert food analyst..."
          }
        },
        "type": "object",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "title": "UsageInfo",
        "description": "Token\u4f7f\u7528\u91cf\u3068\u30b3\u30b9\u30c8\u60c5\u5831"
      },
...
    }
  },
...
}